+++ package.json
@@ -19,7 +19,7 @@
     "axios": "^1.8.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cookies-next": "^5.1.0",
+    "cookies-next": "^4.3.0",
     "lucide-react": "^0.475.0",
     "next": "14.2.23",
     "react": "^18",
